update script to not handle fake images

+++ generate_READMEs.sh
@@ -12,10 +12,14 @@ for dir in *; do
             echo "$dir" | sed "s/./=/g"
             echo
             for image in *.png; do
-                echo "<a href=\"/$ACCOUNT/$REPO/raw/master/$dir/$image\"><img src=\"/$ACCOUNT/$REPO/raw/master/$dir/$image\" width=\"400px\" /></a>"
+                if [ -f $image ]; then
+                    echo "<a href=\"/$ACCOUNT/$REPO/raw/master/$dir/$image\"><img src=\"/$ACCOUNT/$REPO/raw/master/$dir/$image\" width=\"400px\" /></a>"
+                fi
             done
             for image in *.svg; do
-                echo "<embed src=\"/$ACCOUNT/$REPO/raw/master/$dir/$image\" type=\"image/svg+xml\" width=\"400px\" />"
+                if [ -f $image ]; then
+                    echo "<embed src=\"/$ACCOUNT/$REPO/raw/master/$dir/$image\" type=\"image/svg+xml\" width=\"400px\" />"
+                fi
             done
             echo
             ) > $pwd/$dir/README.md
